token값에 맞는 id만 조회,수정,삭제가능

+++ src/controllers/user-controller.js
@@ -1,5 +1,4 @@
 import { userService } from '../services/user-service.js';
-
 export default {
   async createUser(req, res, next) {
     try {
@@ -31,8 +30,18 @@ export default {
   },
 
   async getUser(req, res, next) {
+
+
     try {
       const userId = req.params.userId;
+      if (!req.userId) {
+        return res.status(401).send('Unauthorized');
+      } // 로그인하지 않은 사용자
+  
+      if (req.userId !== userId) {
+        return res.status(403).send('Forbidden');
+      } // 로그인한 사용자와 요청한 사용자가 다른 경우
+  
       const foundUser = await userService.getUser(userId);
       res.status(200).json(foundUser);
     } catch (error) {
@@ -43,6 +52,13 @@ export default {
   async updateUser(req, res, next) {
     try {
       const userId = req.params.userId;
+      if (req.userId !== userId) {
+        res.status(403).json({
+          result: "forbidden-approach",
+          reason: "수정할 권한이 없습니다.",
+        });
+        return;
+      }
       const updatedUser = await userService.updateUser(userId, req.body);
       res.status(200).json(updatedUser);
     } catch (error) {
@@ -53,6 +69,13 @@ export default {
   async deleteUser(req, res, next) {
     try {
       const userId = req.params.userId;
+      if (req.userId !== userId) {
+        res.status(403).json({
+          result: "forbidden-approach",
+          reason: "자신의 계정만 삭제할 수 있습니다.",
+        });
+        return;
+      }
       await userService.deleteUser(userId);
       res.status(204).end();
     } catch (error) {
